clean http stream storage and update sqs
