--- pixels-common/src/main/java/io/pixelsdb/pixels/common/utils/HttpServerHandler.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2023 PixelsDB.
- *
- * This file is part of Pixels.
- *
- * Pixels is free software: you can redistribute it and/or modify
- * it under the terms of the Affero GNU General Public License as
- * published by the Free Software Foundation, either version 3 of
- * the License, or (at your option) any later version.
- *
- * Pixels is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * Affero GNU General Public License for more details.
- *
- * You should have received a copy of the Affero GNU General Public
- * License along with Pixels.  If not, see
- * <https://www.gnu.org/licenses/>.
- */
-package io.pixelsdb.pixels.common.utils;
-
-import io.netty.channel.ChannelHandler;
-
-/**
- * @author jasha64, hank
- * @create 2023-07-27
- * @update 2025-09-18 make this class an interface and clean other methods
- */
-public interface HttpServerHandler extends ChannelHandler
-{
-    void setServerCloser(Runnable serverCloser);
-}
