--- pixels-common/src/main/java/io/pixelsdb/pixels/common/utils/HttpServerUtil.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2023 PixelsDB.
- *
- * This file is part of Pixels.
- *
- * Pixels is free software: you can redistribute it and/or modify
- * it under the terms of the Affero GNU General Public License as
- * published by the Free Software Foundation, either version 3 of
- * the License, or (at your option) any later version.
- *
- * Pixels is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * Affero GNU General Public License for more details.
- *
- * You should have received a copy of the Affero GNU General Public
- * License along with Pixels.  If not, see
- * <https://www.gnu.org/licenses/>.
- */
-package io.pixelsdb.pixels.common.utils;
-
-import io.netty.handler.ssl.SslContext;
-import io.netty.handler.ssl.SslContextBuilder;
-import io.netty.handler.ssl.util.SelfSignedCertificate;
-
-import javax.net.ssl.SSLException;
-import java.security.cert.CertificateException;
-
-/**
- * Some useful methods for server side.
- */
-public final class HttpServerUtil
-{
-
-    private static final boolean SSL = System.getProperty("ssl") != null;
-
-    private HttpServerUtil()
-    {
-    }
-
-    public static SslContext buildSslContext() throws CertificateException, SSLException
-    {
-        if (!SSL)
-        {
-            return null;
-        }
-        SelfSignedCertificate ssc = new SelfSignedCertificate();
-        return SslContextBuilder.forServer(ssc.certificate(), ssc.privateKey()).build();
-    }
-}
