+++ pixels-storage/pixels-storage-http-stream/src/main/java/io/pixelsdb/pixels/storage/http/io/HttpServerInitializer.java
@@ -17,7 +17,7 @@
  * License along with Pixels.  If not, see
  * <https://www.gnu.org/licenses/>.
  */
-package io.pixelsdb.pixels.common.utils;
+package io.pixelsdb.pixels.storage.http.io;
 
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelInitializer;
@@ -28,6 +28,7 @@ import io.netty.handler.codec.http.HttpObjectAggregator;
 import io.netty.handler.codec.http.HttpServerCodec;
 import io.netty.handler.codec.http.HttpServerExpectContinueHandler;
 import io.netty.handler.ssl.SslContext;
+import io.pixelsdb.pixels.common.utils.ConfigFactory;
 
 /**
  * @author jasha64
@@ -35,7 +36,6 @@ import io.netty.handler.ssl.SslContext;
  */
 public class HttpServerInitializer extends ChannelInitializer<SocketChannel>
 {
-
     private final SslContext sslCtx;
     private final ChannelHandler handler;
