+++ pixels-storage/pixels-storage-sqs/pom.xml
@@ -16,8 +16,6 @@
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-
-        <dep.aws.sqs.extend.lib.version>2.1.2</dep.aws.sqs.extend.lib.version>
     </properties>
 
     <dependencies>
@@ -28,16 +26,19 @@
         </dependency>
 
         <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <optional>true</optional>
+            <groupId>io.pixelsdb</groupId>
+            <artifactId>pixels-storage-s3</artifactId>
         </dependency>
 
         <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>amazon-sqs-java-extended-client-lib</artifactId>
-            <version>${dep.aws.sqs.extend.lib.version}</version>
-            <type>jar</type>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>sqs</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+            <optional>true</optional>
         </dependency>
     </dependencies>
